requirements.txt
Django==1.9.5
django-crispy-forms==1.6.0
django-environ==0.4.0
django-extensions==1.6.1
django-filter==0.13.0
django-material==0.8.0
django-rest-swagger==0.3.5
django-super-inlines==0.1.4
djangorestframework==3.3.3
gunicorn==19.4.5
hashids==1.1.0
Pillow==3.2.0
psycopg2==2.6.1
PyYAML==3.11
six==1.10.0
